test-requirements.txt
# vcrpy  # Needed by payment_pagseguro
odoo-test-helper  # Needed by spec_driven_model
pyopenssl==22.1.0
signxml<3.1.0
xmldiff
